MNT: move deprecated xarray functionality into wradlib.io.xarray_depr.py
